+++ packages/components/atoms/f-error-message/test-utils/component-objects/f-error-message.component.js
@@ -1,0 +1,4 @@
+const errorMessageComponent = () => $('[data-test-id="error-message-component"]');
+
+exports.isErrorMessageComponentDisplayed = () => errorMessageComponent().isDisplayed();
+exports.waitForErrorMessage = () => errorMessageComponent().waitForExist();