+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -49,13 +49,6 @@
                 <user-note data-test-id="user-note" />
 
                 <f-button
-                    :class="$style['c-checkout-allergyButton']"
-                    button-type="link"
-                    data-test-id="allergy-button">
-                    {{ $t('allergyText') }}
-                </f-button>
-
-                <f-button
                     :class="$style['c-checkout-submitButton']"
                     button-type="primary"
                     button-size="large"
@@ -480,10 +473,6 @@ $checkout-padding                         : spacing(x5) 100px;
     .c-checkout-alert {
         width: $checkout-width;
         margin: 0 auto;
-    }
-
-    .c-checkout-allergyButton {
-        margin-top: spacing(x1.5);
     }
 
     .c-checkout-submitButton {