+++ packages/components/organisms/f-checkout/src/tenants/en-GB.js
@@ -37,8 +37,6 @@ const messages = {
 
     buttonText: 'Go to payment',
 
-    allergyText: 'If you or someone you’re ordering for has a food allergy or intolerance, click here',
-
     guest: {
         firstName: 'First name',
         lastName: 'Last name',