+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,9 +3,9 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v0.36.0
+v0.37.0
 -------------------------------
-*January 12, 2021*
+*January 14, 2021*
 
 ### Added
 - `Guest`, `Header` and `TermsAndConditions` components.
@@ -13,6 +13,14 @@ v0.36.0
 ### Changed
 - `Checkout` form method.
 - `Checkout` title and login link into `Header` component.
+
+
+v0.36.0
+-------------------------------
+*January 12, 2021*
+
+### Removed
+- Allergen information link.
 
 
 v0.35.0